Fix SonarQube issues in dircache and lantest_dircache_stats
- Remove TOCTOU race condition by eliminating redundant access() check
- Add NOSONAR comments for safe strlen() calls with NULL check justification
- Fix const qualifier warning by creating working copy of buffer
- Fix precision loss in dircache.c by casting both operands to double
- Addresses multiple SonarQube security and code quality warnings

+++ etc/afpd/dircache.c
@@ -622,7 +622,8 @@ void log_dircache_stat(void)
     double hit_ratio = 0.0;
 
     if (dircache_stat.lookups > 0) {
-        hit_ratio = (double)dircache_stat.hits / dircache_stat.lookups * 100.0;
+        hit_ratio = ((double)dircache_stat.hits / (double)dircache_stat.lookups) *
+                    100.0;
     }
 
     /* Get username from AFPobj if available */
